backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9,<0.1
redis>=5.0,<6.0
rq>=1.15,<2.0
python-dotenv>=1.0,<2.0
pydantic>=2.5,<3.0
